fix(meet): authorize shared user settings API sessions
