+++ apps/backend/src/contact/current_user_session_targets.rs
@@ -1,0 +1,21 @@
+pub(super) const CURRENT_USER_APP_SESSION_TARGETS: [&str; 19] = [
+    "calendar",
+    "chat",
+    "cms",
+    "contacts",
+    "drive",
+    "finance",
+    "hive",
+    "inventory",
+    "learn",
+    "mail",
+    "meet",
+    "mind",
+    "mira",
+    "nova",
+    "rewise",
+    "tasks",
+    "teach",
+    "track",
+    "platform",
+];
